+++ src/client/cypress/e2e/editor/coordinates.cy.js
@@ -1,21 +1,7 @@
-import {
-  interceptApiCalls,
-  deleteBorehole,
-  newEditableBorehole,
-  delayedType,
-  login,
-} from "../testHelpers";
+import { newEditableBorehole, delayedType } from "../testHelpers";
 
 describe("Tests for editing coordinates of a borehole.", () => {
   beforeEach(() => {
-    interceptApiCalls();
-
-    login();
-
-    // go to edit
-    cy.get('[data-cy="menu"]').click();
-    cy.contains("h4", "Editor").click();
-    cy.wait("@edit_list");
     newEditableBorehole().as("borehole_id");
 
     cy.get('[data-cy="LV95X"]').as("LV95X-input");
@@ -25,11 +11,6 @@ describe("Tests for editing coordinates of a borehole.", () => {
     cy.get('[data-cy="country"] > input').as("country");
     cy.get('[data-cy="canton"] > input').as("canton");
     cy.get('[data-cy="municipality"] > input').as("municipality");
-  });
-
-  afterEach(() => {
-    // delete borehole
-    cy.get("@borehole_id").then(id => deleteBorehole(id));
   });
 
   it("creates new borehole and adds coordinates", () => {